+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/schema/types/TypeKind.java
@@ -27,5 +27,7 @@ public enum TypeKind {
     ENUM,
     NON_NULL,
     LIST,
-    UNION
+    UNION,
+    INTERFACE,
+    INPUT_OBJECT
 }